Make new item buttons functional
Now each button triggers a form so users can enter information
about the new todo. The dailies dialog is different from the
other list dialog because a) it's difficult to determine from the
submit button which list triggered the add request if everyone
shares the same dialog and b) I'd like to support different
options for dailies than for other todos (namely, repeating).

The click handlers for the submit buttons moved up to the
initialization so multiple handlers wouldn't get added when e.g.
creating a new list or swapping between lists.