Add 'schema' for storing lists in local storage + intro functions.
This commit includes the function declarations for everything I
think I'll need, but only DEFINES the easy functions. It also
moves the object parsing logic out of the todo-classes file.

I decided against storing the todos separately and referencing
them because it would be a lot of trouble to manually stringify
the list metadata EXCLUDING the todos (only having their IDs), and
the only list metadata separate from the todos is the name -- as
the "next todo ID" field needs to be updated as well.

If this were a production app, it would be worth it: I'd have a
more robust way to generate unique IDs for todos and lists, and
the list name and list of TODOs would be stored separately...
but in a production app I wouldn't be writing to this single
bucket local storage table.