Stores lists and todos to local storage.
This still retains some helper methods for development. I'm aware
that dom-manager really should not touch anything in
storage-manager. For a project of this size, I'm completely
comfortable directly reading the color theme there instead of
passing it through from index.js (which seems excessive). But the
inclusion (in this change) of the on-todo-completed write to
local storage? Gross. I know I can pass a callback into dom-manager
somewhere, maybe even make it a class that gets initialized. But
this is okay _enough_ for right now.