Add logic to change list name when clicking edit button
This change is one out of several changes I intend to make when
swapping to "edit mode" when clicking the edit button. Next up is
adding a delete button for each todo, but this requires
refactoring the todo arrays into objects -- this change adds the UI
but does not add logic to actually SHOW these buttons or make them
functional.