Fix icons moving when label is replaced with text input
Since the auto margin was on the label, the icons would shift over
when the label was replaced with the text input in edit mode. I
chose not to remove the auto margin on the label since it also
applies to other items (label on the list sidebar, dailies) that
might use it later.